==> 02_Pre_Analysis.Rmd <==
---
title: "Tech 2. Pre-Analysis 5% Sample."
output:
  html_document
---

This is the analysis carried out in the 'Data Finder' phase of the SCORE project used for the pre-registration.

```{r setup, include=FALSE}
pacman::p_load("tidyverse","kableExtra","MASS","sjPlot","webshot","knitr")
```


```{r load, warning = F, message = F}
df_orig_complete <- read_csv(file = here::here("data/df_orig_complete.csv"))
```



Here I select out 5% of the sample at random and run their models as part of the pre-registration.

```{r sample5, warning = F, message = T}

set.seed(90210)

df_orig_complete <- df_orig_complete %>%
  mutate(rand = runif(NROW(df_orig_complete$y1)),
         p5 = ifelse(rand > 0.05, NA, 1))

df_orig_complete_5 <- df_orig_complete[!is.na(df_orig_complete["p5"]),]



```


```{r tbl5_rep_pre, warning = F, message = F}
Tbl5_m1 <- polr(factor(stdliv_next5) ~ factor(cat_age) + factor(wave) + factor(cat_age)*factor(wave), data = df_orig_complete_5, Hess = T)

Tbl5_m2 <- polr(factor(stdliv_next5) ~ factor(cat_age) + factor(wave) + factor(cat_age)*factor(wave) + female + factor(education) + factor(EGP6) + factor(income), data = df_orig_complete_5, Hess = T)

Tbl5_m3 <- polr(factor(stdliv_next5) ~ factor(cat_age) + factor(wave) + factor(cat_age)*factor(wave) + female + factor(education) + factor(EGP6) + factor(income) + pensions + unemployed + car_owner + factor(cntry), data = df_orig_complete_5, Hess = T)

tab_model(Tbl5_m1, Tbl5_m2, Tbl5_m3, p.style = "stars", show.ci = F, show.loglik = T,  pred.labels = c("k1","k2","k3","k4","Age 30-44","Age 44-59","Age >60","Year 2007","Age 30-44*year'07", "Age 44-59*year'07", "Age >60*year'07", "Female", "Educ mid", "Educ high" ,"EGP: routine non-man", "EGP: Self", "EGP: Skilled", "EGP: Unskilled", "EGP: Farmers", "EGP: Never had a job", "Income mid", "Income high","Income missing", "Pensions and benefits","Unemployed", "Car","bulgaria", "czech", "estonia", "hungary","latvia", "lithuania", "moldova","poland", "romania", "russia","slovakia", "ukraine"), file = "results/Tbl5_rep.htm")

```

Output Table

```{r tbl5_rep_preo, warning = F, message = F}
#not used in the end

#webshot("results/Tbl5_rep.htm", file = "results/Tbl5_rep.png")

#knitr::include_graphics(here::here("results/Tbl5_rep.png"))
```

### Colophon

```{r}
sessionInfo()
```
